Move geoip code to seperate module
